printk: external log buffer (CONFIG_LOGBUFFER)
When enabled via CONFIG_LOGBUFFER, this feature allows a bootloader to
pass a buffer with existing log messages to the kernel via an optional
command line parameter, which ensures cross arch portability.  The
parameter specifies the address of a log control block which is used to
replace the default log buffer.  Existing bootloader and kernel log
messages are kept in order inside the new buffer.  After a warm boot,
this allows a bootloader to display bootloader and kernel log entries
from previous boots and allows the kernel to display bootloader log
entries along with its own messages. Also, it allows log entries from
multiple boots to be visible in the log to both bootloader and kernel.

This feature is intended for debug purposes and has no effect unless the
command line parameter is specified.  Further, it validates the passed
control block carefully and if any checks fail, it falls back to the default
behaviour.  As such, it can be left enabled by default.

Memory Reservation
------------------
This feature expects the bootloader to reserve the memory using a separate
mechanism.  Reserving the memory prevents the kernel from inadvertently
stomping the external log control block and log entries.  For the greatest
flexibility, I recommend using the mem or memmap kernel command line args
to reserve the memory from within the bootloader.  However, using the
devicetree should work as well.

Based on the initial work of Wolfgang Denk and Igor Lisitsin [1].
Also based on work by: Alexander Streit <Alexander.Streit@continental-corporation.com>

[1] http://git.denx.de/?p=linux-2.6-denx.git;a=commitdiff;h=212f61c7fd3b952a81d1459dd32a86a32ddfd4ce

Signed-off-by: Sean Hudson <sean_hudson@mentor.com>
Reviewed-by: Wade Farnsworth <wade_farnsworth@mentor.com>
Reviewed-by: Joe MacDonald <joe_macdonald@mentor.com>